requirements.txt
streamlit
pandas
fpdf
python-docx